requirements.txt
django-debug-toolbar==1.3.2
pluggy==0.3.0
py==1.4.30
sqlparse==0.1.16
tox==2.1.1
